Move RenderMessageMarkdown to sparkle
