+++ front/components/actions/tables_query/TablesQueryActionDetails.tsx
@@ -11,11 +11,9 @@ import { useCallback, useContext } from "react";
 
 import { ActionDetailsWrapper } from "@app/components/actions/ActionDetailsWrapper";
 import type { ActionDetailsComponentBaseProps } from "@app/components/actions/types";
-import {
-  CodeBlock,
-  RenderMessageMarkdown,
-} from "@app/components/assistant/RenderMessageMarkdown";
-import { ContentBlockWrapper } from "@app/components/misc/ContentBlockWrapper";
+import { ContentBlockWrapper } from "@app/components/assistant/markdown/ContentBlockWrapper";
+import { RenderMessageMarkdown } from "@app/components/assistant/markdown/RenderMessageMarkdown";
+import { CodeBlock } from "@app/components/markdown/CodeBlock";
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 
 export function TablesQueryActionDetails({
