+++ front/components/actions/websearch/utils.ts
@@ -1,5 +1,5 @@
 import type { Citation } from "@dust-tt/sparkle";
-import type { WebsearchActionType } from "@dust-tt/types";
+import type { WebsearchActionType, WebsearchResultType } from "@dust-tt/types";
 
 interface WebsearchResultCitation {
   href: string;
@@ -7,17 +7,17 @@ interface WebsearchResultCitation {
   type: Exclude<React.ComponentProps<typeof Citation>["type"], undefined>;
 }
 
+export function makeWebsearchResultsCitation(result: WebsearchResultType) {
+  return {
+    description: result.snippet,
+    href: result.link,
+    title: result.title,
+    type: "document" as const,
+  };
+}
+
 export function makeWebsearchResultsCitations(
   action: WebsearchActionType
 ): WebsearchResultCitation[] {
-  return (
-    action.output?.results.map((r) => {
-      return {
-        description: r.snippet,
-        href: r.link,
-        title: r.title,
-        type: "document",
-      };
-    }) ?? []
-  );
+  return action.output?.results.map(makeWebsearchResultsCitation) ?? [];
 }
