+++ front/components/assistant/markdown/MarkdownCitation.tsx
@@ -1,0 +1,7 @@
+import type { Citation } from "@dust-tt/sparkle";
+
+export interface MarkdownCitation {
+  href?: string;
+  title: string;
+  type: Exclude<React.ComponentProps<typeof Citation>["type"], undefined>;
+}
